Removed the `reset' functions since the run-time is doing an automatic reset which set the values for storing and retrieving to their previous value after a store/retrieve operation.
So now, we call the `init' functions at the very begining of the launch of EiffelBench
and then the run-time is doing the job of `reset'.

Remove an arguments to `retrieve_all' and `partial_retrieve'. We will maybe add it
again later, if we find a way to specify the size we need to read.

SVN: 10296@trunk