Improvement on the speed of the finalization time.
First, I removed all the *_UNIT* classes from the `polymorphism' cluster. I merged the definition
with the *_ENTRY* classes. Why? Basically, because from the *_UNI* which were stored in the EIFGEN
directory we were rebuilding a set of *_ENTRY* classes. This implies a big penalty, because we
are using SORTED_TWO_WAY_LIST as a layout structure and each time we are inserting an element it takes
forever.

The first step was to switch all the code from *_UNIT* into *_ENTRY* and this has been done without
any problems. However, we thought it was not enough because the data structure was not fast enough,
so we decide to change it and to use a sorted array.

Since the creation of the data structures is done mostly at degree -4, we decided to create unsorted
array with elements and just before to store them in the EIFGEN we will sort them using the quicksort
algorithm. The change in the data structure implies to update all the traversal and since it was
sorted I've created a binary search algorithm to do a quick search on the data structure.

Now, it is very fast to finalize.

SVN: 10419@trunk