In the previous version, we were saying that the `poofter_finalization' was useless, but I was wrong. The problem is that we cannot be sure after some incremental compilations that the order in the assigned CLASS_TYPE_IDs will be the same as the one given by the topological sort.
So when doing the poofter_finalization we just need to compare the conformance between
two items trough the conformance table of the found CLASS_C. Otherwise, we need to
use `simple_conform_to' to be sure that the conformance test is correct.

For ATTR_TABLE, we need to find first the offset position of the sought attribute,
which means that we are not very efficient, but we at least produce a very efficient
code.

SVN: 11877@trunk