Update `Precursor' to follow the new rules which are:
        A Precursor is valid if and only if it satisfies the
        following five conditions:

            1. It appears in the Routine_body or the Rescue_clause
            of the declaration of exactly one routine `r' of C.

            2. If the Parent_qualification part is present, its
            Class_name is the name of a parent P of C.

            3. The declaration of `r' in C is the redefinition of a
            routine that has exactly one effective version in the set
            of Parent types of C, restricted if case 2 applies to
            those having P as their Class_name.

Thus we removed VUPR4 and VUPR5 from the system.

There was also a problem due to the use of Precursor in the case where the
parents are of a generic type, in this case `Precursor' should be forbidden
but was not because there was the same TYPE_A and the same ROUTINE_ID, because
the storage was done through an HASH_TABLE indexed by ROUTINE_ID. Now,
it is a LINKED_LIST of PAIR [CL_TYPE_A, ROUTINE_ID]].

This also solve the case of the multiple and repeated inheritance of a feature
in a class A where we want to redefine it by using the Precursor keyword. Before
it was allowed not to specify the parent, but this was incorrect. The problem was
linked to the HASH_TABLE problem described above.

SVN: 12409@trunk