Fixed a bug with incremental finalization using `inlining': Starting from a compilation of 	f is do g end 	g is do end to 	f is do g end 	g is do another_feature do
This change did not result in a correct generation because the inlining of `g' did
not generate the call to `another_feature'.

The problem is that we were using in `f' the body_id of `g', however after the modification
the compiler changed the body_id of `g' into something else, however, `f' was still using
the old body_id which results in an incorrect generation.

The solution was to stored `body_index' instead and to do a look up in the body_index_table
to find the real body_id of `g'.

SVN: 14180@trunk