Add the generation of the `finished' file for the makefile. In most of the case the `finished' file are created by the makefile and removed by the compiler (done here).
Fixed a problem due to the addition of a C++ external to an Eiffel class. The
result was to change the C generated file name from `co23.c' to `co23.cpp'. However
the makefile did not see the changes because the .c file was still older than the .obj
file. The only solution is to remove the .c file which is what we are doing now.

In order to do it, we need to keep an information about the C++ status of a class.
This status is only valid in workbench mode, not in final mode because of inlining.

SVN: 15002@trunk