Used EIF_REFERENCE instead of `char *'.
Fixed a very old bug in the debugger which was easy to fix thanksful to the same
kind of bug found in `rout_obj.c'. Basically we were creating two eiffel strings
without protecting the references. As a result, the first string could have moved
after the creation of the second one and most of the time is was crashing EiffelBench.

SVN: 17024@trunk