Create an empty conformance table at the beginning. It will be created when necessary.
New generation of C files which does not make use of D and F directories. This has been
done to improve the performance of the C compilation, especially when using
`quick_finalize'. Now, all the files corresponding to one class are generated in
the same C directory, therefore when you modify one class you only have to recompile
one C directory and not 3 as it used to be.

SVN: 17555@trunk