Fixed a memory leak problem when a Window which was not yet destroyed was collected, we were not calling `DestroyWindow' on it. Now we do it during the call to `dispose' from WEL_WINDOW through `gc_destroy_item'. Since a call to `DestroyWindow' will send the message `WM_DESTROY' to its children if any before destroying them, we need to free the memory used by OBJECT_ID. Unfortunately we cannot do it from the standard WEL_DISPATCHER because at final dispose the object can be already collected.
So we do it with a special trick in the C routine used to call the WEL_DISPATCHER object,
but before we set the dispatcher object to null so that it goes to the second part of
`cwel_window_procedure', then we catch only the WM_DESTROY message, get the internal user
data and if it is valid (ie > 0) we free the entry in the object id stack and proceed
by calling the standard DefWindowProc. When this is done, we restore the dispatcher object
pointer to its previous value and we continue.

When an object is being disposed and has already been destroyed by Windows, its internal_data
is 0 and we do not do anything.

SVN: 23387@trunk