Fixed memory leak because we were not calling the existing default window procedure of WEL_CONTROL objects when they get distroyed. Now, in the GWL_USERDATA part we have a small C structure EIF_WEL_USERDATA (instead of just having the object id) which contains 2 data: the object_id and the default window procedure. When we received the WM_DESTROY message on the C side, we check if we have a default window procedure, if so we call it, otherwise we call the windows default one.
This structure is either destroyed on the Eiffel side in WEL_WINDOW.destroy_item
when Eiffel receives the WM_DESTROY message, or on the C side when the GC
disposes the Eiffel objects and generates automatically a WM_DESTROY message.

We are now using `GetWindowLongPtr' and `SetWindowLongPtr' for future compatibility.

SVN: 32916@trunk