Improved `is_polymorphic' to find more static calls. The issue is that the algorithm used was solely based on CLASS_C to perform conformance between two elements of current table. However if those two elements are in a different generic derivation it will return True, whereas most of the time it is not.
Fixed bug where `is_polymorphic' did a bad job when there was only one element
in the table, and that this element was just the implementation of a deferred
routine in a generic class (related to the above bug fix too). This fixed issue
`ccomp040' of eweasel regression test.

SVN: 37642@trunk