Fixed a dead lock when two threads (with the same parent thread) are exiting at the same time and a GC synchronization is requested by one. Because the location of the request the `children_mutex' was already taken by the one requesting the GC synchronization, leaving the other thread blocked on trying to acquire the mutex. But the GC synchronization did not know about it and therefore waited indefinitely. Shown by eweasel test exec120. We now mark the thread blocked just before acquiring the mutex.
FIXME: we moved the call to eif_wean before acquiring the mutex because it
can be done at any time and also it removes a potential dead lock if the
call to eif_wean if it triggers a GC synchronization. Moreover there is
a bug as henter/eif_wean are done per thread and not on a globally scope
therefore we insert an incorrect location to the free_list, but it does
not matter as we are quitting the thread. The real issue is that the parent
thread `hec_saved' is not freed of a location although we put a NULL reference
at the location.

SVN: 38517@trunk