Re-implemented `initialize_docking_areas' as it was not initializing the areas correctly while one or more widgets are docked out. Instead of not iterating `all_holders', `linear_represntation' is iterated, and the current holder determined from this index.
To reproduce the bug that this fixed, put three items in a multiple split
area, drag the second out, and try to adjust the position of the first
by dragging. It will let you position the widget in two places, when
really it should only allow the final place to be used for insertion.

SVN: 38613@trunk