No need for PROTECT_B anymore as it was causing eweasel test `exec167' to fail. This failed because the protection was done prior the call to the encapsulation and the assertions usually did not use the protected reference but the unprotected one which did not exist.
Instead now, we generate the call normally and let the inside of the encapsulation
take care of the protection. Therefore when such a protection is needed we
always call the encapsulation, it is less efficient but not that much.

Moreover we now handle well expanded types when part of arguments of an
external.

SVN: 38876@trunk