Fixed bug in main loop of `set_with_rtf'. We were checking for a space to trigger the start of text at the subsequent character before we were checking that the previous character was not a control character.
This meant that the following rtf snippet "} {" would not treat the space
as text, but instead start looking for text from the '{' character. Of course
no text would be found, and the text processing would end immediately. This
caused spaces between formatting sections to be lost in that case.

SVN: 43517@trunk