Two major bug fixes.
1. If a document did not contain a reference to a null color and had
no mention of a higlight color, the highlight color used was the first
color in the color table. This is because character format had no method of
determining if a color had been explicitly set. Added `bcolor_set' to combat
this and within `append_color_for_rtf' we check that `bcolor_set' is True
before appending a background color to the internal rich text.

2. The processing of the fonttable did not work for the following fonttable:
{\fonttbl\f0\froman\fcharset0 Times New Roman;\f1\froman\fcharset0 Times New
Roman;} as the code expected each new font to be enclosed within {} as all
the test files up to this point had. This caused a system to crash if
loading a file with this type of fonttable. We now always add a new font
to `all_fonts' after processing the font name and not only when we find
the '}' character.

SVN: 45121@trunk