Fixed contracts violations, added contracts:
SD_TOOL_BAR_GRAGGING_AGENTS: 

The reason of change already wrote in the source codes.


SD_ZONE:

The reason is when a SD_FLAOTING_ZONE have multiply contents which are docking at sides of each other. Such as a floating zone have Features tool at top half and Clusters tool at bottom half. At this time, when dragging the top title bar of the floating zone, the `SD_ZONE.content? query will be void. This is a safe change, because we already handle void result of SD_ZONE.content if end user dragging a top title bar in the past.


SD_HOT_ZONE_OLD_MAIN

Just added a check contract, I think added a contract at this place is good. It?s useful to find the reasons of bugs such as: bug#13066 and bug#13094.


SD_OPEN_CONFIG_MEDIATOR

We added if protection to test if the splitter position is valid. When the splitter is invisible which caused by not enough space (such as: main development window too small), it well crash here before (at least on GTK). Now we added if protection, it works fine now.

reviewer: Manu

SVN: 69100@trunk