Added if protection to EB_WARNINGS_TOOL, then the precondition `is_displayed' will not be violated.
For the changes in SD_DOCKING_STATE and SD_TAB_STATE, it's for the bug that just after auto hide the whole contexts tools, the Output tool didn't slide out if end user selecting it in the ctrl tab dialog. The reason of this bug is: we have not cleared last focus content value in the past before a tool auto hiding. So when selecting the Output tool in ctrl tab dialog, last focus content still was Output tool (this compare is done in SD_CONTENT.set_focus), then it will not slide out. Now, we clear the value in auto hiding features (`stick' feature) of SD_DOCKING_STATE and SD_TAB_STATE, it will slide out now.

Fixed bug#13101: Violation of precondition `is_displayed' of {EV_DRAWING_AREA}.set_focus

Reviewer: Manu, Ted

SVN: 69137@trunk