Changed how eweasel handles missing test_end instructions and missing newline at end of file in a test control file.
Old behavior:  Missing newline at EOF was allowed but any instruction on that
line was not read.  A missing test_end instruction was allowed but the test
for this condition didn't allow the test_end to be anywhere except the last
line of the file.

New behavior:  A missing test_end instruction is an error.  A test_end instruction
anywhere in the file is allowed - test_end does not have to be the last instruction.
A missing newline at EOF is an error, but any instruction on the last partial line
will be processed before the error is reported.

SVN: 78303@trunk