Now `{EV_SUGGESTION_WINDOW}.build_suggestion_list': * will cause an automatic selection of the closest match so no need to call `select_closest_match' all the time after building that list. * will always resize the popup window to adapt to content (with 2 options: one is to fit the content of the list all the time, i.e. growing and shrinking, the other one is to only grow the window, but never shrink it). The resizing is controled by `{EV_SUGGESTION_SETTINGS}.is_fit_list_window_to_content'.
Added helpers `required_grid_width' and `required_grid_height' which tell us how big should the suggestion window be to display all items without truncation. Fixed a bug in that computation in the event where the required height is too large to fit in the grid, we know that a vertical scrollbar will appear and thus we need to make the width slightly larger otherwise an hortizontal scrollbar would appear.

Fixed a bug in setting the size of the suggestion window where we forgot to take into account the 1 pixel border around the grid which caused scrollbars to appear when we did not want them and that we had set the size to made sure that no scrollbars would appear.

Disable partial dynamic content as it is causing too many problems as many queries on the grid are not going to evaluate the partial dynamic function to create the item.

SVN: 91209@trunk