Added CURLOPT_NOSIGNAL to the Eiffel cURL wrapper's constants.
This could be used on linux system to handle timeout problem like "cURL Error (28): Timeout was reached".

The problem is that on (Li|U)nix, when libcurl uses the standard name resolver, a SIGALRM is raised during name resolution which libcurl thinks is the timeout alarm.

The solution is to disable signals using CURLOPT_NOSIGNAL.

SVN: 93000@trunk