Introduced a new public API.
The new API mainly lives in TRANSACTION_CONTEXT and doesn't expose
as many features as the old EXECUTOR. For example the delete operation
and write operations outside a transaction have been removed.

The idea is to never let the user manually delete an object, as it is 
potentially dangerous when the deleted object is shared. Instead
of manual deletion a garbage collector should reclaim unused objects.

To implement a garbage collector however the new API introduces the 
concept of a 'root object'. These objects and all referenced objects
will never be reclaimed. By default the top of an object graph that 
gets inserted is a 'root object', but the user can change this default
or set the root annotation manually as well.

Furthermore the funcionality to execute queries within implicit
transactions has been moved to the REPOSITORY, and QUERY has
gained a function to filter objects based on the root status.

SVN: 93231@trunk