Further performance improvements:
*) Added two new functions in READ_MANAGER which should avoid
some unnecessary and costly checks. The price is a violation
of the command-query separation, but I think in this case it's
acceptable.

*) Removed {PS_OBJECT_READ_DATA}.to_initialize. The reduced memory
usage is worth more than avoiding to recompute the type.

*) Added an attribute to INTERNAL_TRANSACTION to store the associated
identifier set. This avoids an expensive lookup.

The set of patches can reduce execution time by about 30%.

SVN: 93760@trunk