Remove setting of background color to white to avoid white selection too
The GTK 3 code for setting the background color is using a GTK 3.0 API that should not be used anymore. While we are still using it, looks like there is no need to set the background to white in the initialization of the EV_TEXT_IMP, as this is not done in the Windows version too.

SVN: 99092@trunk