Fix crash after flush_buffer
Because we forgot to keep a reference of `text_buffer` when getting it from GTK, after calling `flush_buffer` any attempts to use the new buffer would crash. Not sure why it was not already failing in GTK 2.0.

This allows the rich_edit sample to load the `welcome.rtf` file.

SVN: 99095@trunk