Updated code to support resizing (to bigger and also to smaller value) the SD auto hize zone.
Take into account the panel of the SD_AUTO_HIDE_ZONE, and not the full zone that has a minimum size (on gtk3) due to EV_FIXED.set_item_size ... implementation on gtk3 requiring to use gtk_widget_set_size_request).
Also take into account the minimum size when ending a resize operation, only if it was set on the interface  (i.e not the real minimum size set by the implementation).

SVN: 105880@trunk